Add more optimizations - Tried many small changes to attempt to reduce register use but was   unable to meet the 32 register target that would allow a full launch.   Adding __launch_bounds__ with max 256 threads min 8 blocks forced   compilation to use only 32 registers without any spill over netting   a 10% run time reduction - Replaced the iterative guess and check method of getting random unit   vectors with a method using projection of a sphere onto a cylinder - At this point memory dependency became by far the dominant reason   for stalls. Attempted to make thread access of sphere data sequential   rather than simultaneous but this did not have the desired effect. - Run time to render the full sample/bounce cover image is now around   33s compared with 63s at the start, nearly 2x improvement in speed

+++ RayTracer/src/vec3.cuh
@@ -81,12 +81,12 @@ namespace rtw
 
 		__host__ __device__ float length() const
 		{
-			return std::sqrt(length_squared());
+			return sqrtf(length_squared());
 		}
 
 		__host__ __device__ Vec3& normalize()
 		{
-			*this *= (1.0f / std::sqrt(length_squared()));
+			*this *= (1.0f / sqrtf(length_squared()));
 			return *this;
 		}
